Move TunerSwerveDrivetrain to new module
